Convert it to monorepo
