+++ .gitignore
@@ -3,3 +3,4 @@
 # dependencies
 backend/node_modules
 frontend/node_modules
+node_modules
