+++ apps/frontend/public/manifest.json
