+++ apps/frontend/public/robots.txt
