+++ packages/eslint-config-acme/index.js
@@ -1,0 +1,7 @@
+module.exports = {
+  extends: ["next", "prettier"],
+  rules: {
+    "@next/next/no-html-link-for-pages": "off",
+    "react/jsx-key": "off",
+  },
+};
